% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/app_RLum.R
\name{app_RLum}
\alias{app_RLum}
\title{Run Luminescence shiny apps}
\usage{
app_RLum(app, ...)
}
\arguments{
\item{app}{\code{\link{character}} (required): name of the RLum app to start. See details for a list
of available apps.}

\item{...}{further arguments to pass to \code{\link{runApp}}}
}
\description{
A wrapper for \code{\link{runApp}} to start interactive shiny apps for the R package Luminescence.
}
\details{
\tabular{lcl}{
\bold{Application name:} \tab    \tab \bold{Function:} \cr
abanico \tab -> \tab \code{\link{plot_AbanicoPlot}} \cr
histogram \tab -> \tab \code{\link{plot_Histogram}} \cr
KDE \tab -> \tab \code{\link{plot_KDE}} \cr
radialplot \tab -> \tab \code{\link{plot_RadialPlot}} \cr
doserecovery \tab -> \tab \code{\link{plot_DRTResults}} \cr
cosmicdose \tab ->  \tab \code{\link{calc_CosmicDoseRate}}
}
}
\examples{
\dontrun{
# Plotting apps
app_RLum("abanico")
app_RLum("histogram")
app_RLum("KDE")
app_RLum("radialplot")
app_RLum("doserecovery")

# Further apps
app_RLum("cosmicdose")
}
}
\author{
Christoph Burow, University of Cologne (Germany)
}
\seealso{
\code{\link{runApp}}
}

