% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLumShiny.R
\docType{package}
\name{RLumShiny-package}
\alias{RLumShiny}
\alias{RLumShiny-package}
\title{Shiny Applications for the R Package Luminescence}
\description{
A collection of shiny applications for the R package Luminescence.
These mainly, but not exclusively, include applications for plotting chronometric
data from e.g. luminescence or radiocarbon dating. It further provides access to
bootstraps tooltip and popover functionality as well as a binding to JSColor.
}
\details{
In addition to its main purpose of providing convenient access to the Luminescence
shiny applications (see \code{\link{app_RLum}}) this package also provides further functions to extend the
functionality of shiny. From the Bootstrap framework the JavaScript tooltip and popover
components can be added to any shiny application via \code{\link{tooltip}} and \code{\link{popover}}.
It further provides a custom input binding to the JavaScript/HTML color picker JSColor.
Offering access to most options provided by the JSColor API the function \code{\link{jscolorInput}}
is easily implemented in a shiny app. RGB colors are returned as hex values and can be
directly used in R's base plotting functions without the need of any format conversion.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tzerk.github.io/RLumShiny/}
  \item Report bugs at \url{https://github.com/tzerk/RLumShiny/issues}
}

}
\author{
\strong{Maintainer}: Christoph Burow \email{christoph.burow@gmx.net} (\href{https://orcid.org/0000-0002-5023-4046}{ORCID})

Authors:
\itemize{
  \item Christoph Burow \email{christoph.burow@gmx.net} (\href{https://orcid.org/0000-0002-5023-4046}{ORCID})
  \item Urs Tilmann Wolpert
  \item Sebastian Kreutzer (\href{https://orcid.org/0000-0002-0734-2199}{ORCID})
  \item Marco Colombo (\href{https://orcid.org/0000-0001-6672-0623}{ORCID})
}

Other contributors:
\itemize{
  \item  R Luminescence Package Team [contributor]
  \item Jan Odvarko (jscolor.js in www/jscolor) [copyright holder]
  \item  AnalytixWare (ShinySky package) [copyright holder]
  \item  RStudio (chooser_inputBinding.js in www/ and chooser.R in R/) [copyright holder]
}

}
